feat: account model

+++ SU24SE069_PLATFORM_KAROKE_DataAccess/Models/Item.cs
@@ -10,7 +10,7 @@ namespace SU24SE069_PLATFORM_KAROKE_DataAccess.Models
         public string ItemName { get; set; } = null!;
         public string ItemDescription { get; set; } = null!;
         public int ItemType { get; set; }
-        public double ItemPrice { get; set; }
+        public decimal ItemPrice { get; set; }
         public int ItemStatus { get; set; }
         public int CanExpire { get; set; }
         public int CanStack { get; set; }
